+++ src/components/lucide/ArrowLeft.vue
@@ -1,0 +1,44 @@
+<template>
+    <svg xmlns="http://www.w3.org/2000/svg" 
+        :width="props.size" 
+        :height="props.size" 
+        viewBox="0 0 24 24" 
+        fill="none" 
+        stroke="currentColor" 
+        :stroke-width="props.strokeWidth" 
+        stroke-linecap="round" 
+        stroke-linejoin="round" 
+        class="lucide lucide-arrow-left">
+        <path d="m12 19-7-7 7-7"/>
+        <path d="M19 12H5"/>
+    </svg>
+</template>
+
+<script lang="ts">
+/**
+ * We integrate the required LucideIcons manually, to reduce the build effort and not setting any 
+ * requirements for the icon package.
+ * @source https://lucide.dev/icons/arrow-left
+ */
+export interface LucideIconProps {
+    size?: 24 | number;
+    strokeWidth?: number | string;
+    absoluteStrokeWidth?: boolean;
+}
+
+// Default Export, used for IDE-related auto-import features
+export default {
+    name: 'ArrowLeft'
+}
+</script>
+
+<script lang="ts" setup>
+// Define Component
+const props = withDefaults(defineProps<LucideIconProps>(), {
+    size: 24,
+    strokeWidth: 1.5,
+    absoluteStrokeWidth: false
+});
+</script>
+
+<style scoped></style>